HL2SDK: resolve include conflicts & fix errors
