+++ public/tier1/utlstring.h
@@ -10,10 +10,10 @@
 #pragma once
 #endif
 
-#include "tier1/utlstringtoken.h"
+#include <limits.h>
+#include "strtools.h"
+#include "tier1/utlcommon.h"
 #include "tier1/utlmemory.h"
-#include "tier1/strtools.h"
-#include "limits.h"
 
 #if defined( OSX )
 inline wchar_t *wcsdup(const wchar_t *pString)
@@ -148,8 +148,8 @@ public:
 	DLL_CLASS_IMPORT static CUtlString PathJoin(const char *pStr1, const char *pStr2);
 
 	// These can be used for utlvector sorts.
-	static int __cdecl SortCaseInsensitive(const CUtlString *pString1, const CUtlString *pString2);
-	static int __cdecl SortCaseSensitive(const CUtlString *pString1, const CUtlString *pString2);
+	static int SortCaseInsensitive(const CUtlString *pString1, const CUtlString *pString2);
+	static int SortCaseSensitive(const CUtlString *pString1, const CUtlString *pString2);
 
 	DLL_CLASS_IMPORT void Purge();
 	void Clear() { Purge(); }
@@ -303,7 +303,7 @@ inline const char *CUtlString::Get() const
 {
 	if(!m_pString)
 	{
-		return "";
+		return StringFuncs<char>::EmptyString();
 	}
 	return m_pString;
 }
@@ -322,66 +322,17 @@ inline bool CUtlString::IsEmpty() const
 	return !m_pString || m_pString[0] == 0;
 }
 
-inline int __cdecl CUtlString::SortCaseInsensitive( const CUtlString *pString1, const CUtlString *pString2 )
+inline int CUtlString::SortCaseInsensitive( const CUtlString *pString1, const CUtlString *pString2 )
 {
 	return V_stricmp( pString1->String(), pString2->String() );
 }
 
-inline int __cdecl CUtlString::SortCaseSensitive( const CUtlString *pString1, const CUtlString *pString2 )
+inline int CUtlString::SortCaseSensitive( const CUtlString *pString1, const CUtlString *pString2 )
 {
 	return V_strcmp( pString1->String(), pString2->String() );
 }
 
 
-
-//-----------------------------------------------------------------------------
-// Purpose: Implementation of low-level string functionality for character types.
-//-----------------------------------------------------------------------------
-
-template < typename T >
-class StringFuncs
-{
-public:
-	static T		*Duplicate( const T *pValue );
-	// Note that this function takes a character count, and does not guarantee null-termination.
-	static void		 Copy( T *out_pOut, const T *pIn, int iLengthInChars );
-	static int		 Compare( const T *pLhs, const T *pRhs );
-	static int		 CaselessCompare( const T *pLhs, const T *pRhs );
-	static int		 Length( const T *pValue );
-	static const T  *FindChar( const T *pStr, const T cSearch );
-	static const T	*EmptyString();
-	static const T	*NullDebugString();
-};
-
-template < >
-class StringFuncs<char>
-{
-public:
-	static char		  *Duplicate( const char *pValue ) { return strdup( pValue ); }
-	// Note that this function takes a character count, and does not guarantee null-termination.
-	static void		   Copy( char *out_pOut, const char *pIn, int iLengthInChars ) { strncpy( out_pOut, pIn, iLengthInChars ); }
-	static int		   Compare( const char *pLhs, const char *pRhs ) { return strcmp( pLhs, pRhs ); }
-	static int		   CaselessCompare( const char *pLhs, const char *pRhs ) { return Q_strcasecmp( pLhs, pRhs ); }
-	static int		   Length( const char *pValue ) { return (int)strlen( pValue ); }
-	static const char *FindChar( const char *pStr, const char cSearch ) { return strchr( pStr, cSearch ); }
-	static const char *EmptyString() { return ""; }
-	static const char *NullDebugString() { return "(null)"; }
-};
-
-template < >
-class StringFuncs<wchar_t>
-{
-public:
-	static wchar_t		 *Duplicate( const wchar_t *pValue ) { return wcsdup( pValue ); }
-	// Note that this function takes a character count, and does not guarantee null-termination.
-	static void			  Copy( wchar_t *out_pOut, const wchar_t  *pIn, int iLengthInChars ) { wcsncpy( out_pOut, pIn, iLengthInChars ); }
-	static int			  Compare( const wchar_t *pLhs, const wchar_t *pRhs ) { return wcscmp( pLhs, pRhs ); }
-	static int			  CaselessCompare( const wchar_t *pLhs, const wchar_t *pRhs ); // no implementation?
-	static int			  Length( const wchar_t *pValue ) { return (int)wcslen( pValue ); }
-	static const wchar_t *FindChar( const wchar_t *pStr, const wchar_t cSearch ) { return wcschr( pStr, cSearch ); }
-	static const wchar_t *EmptyString() { return L""; }
-	static const wchar_t *NullDebugString() { return L"(null)"; }
-};
 
 //-----------------------------------------------------------------------------
 // Dirt-basic auto-release string class. Not intended for manipulation,
