+++ public/tier1/utlsymbol.h
@@ -207,7 +207,7 @@ public:
 	};
 
 	typedef CUtlHashtable<UtlSymElm_t, empty_t, UtlSymTableHashFunctor, UtlSymTableEqualFunctor, UtlSymTableAltKey> Hashtable_t;
-	typedef CUtlVector<MemBlockHandle_t, UtlSymElm_t> MemBlocksVec_t;
+	typedef CUtlVector<MemBlockHandle_t> MemBlocksVec_t;
 
 	const Hashtable_t &GetHashtable() const
 	{
